+++ examples/example/App.tsx
@@ -7,9 +7,11 @@ import {
   DripsyProvider,
   Container,
   Theme,
+  Pressable,
 } from 'dripsy'
 // Import from core
 import { H4 } from '@dripsy/core'
+import Gradient from '@dripsy/gradient'
 import { Text } from 'react-native'
 
 const theme = {
@@ -28,6 +30,11 @@ const theme = {
   text: {
     primary: {
       fontSize: 40,
+      color: 'green',
+    },
+    secondary: {
+      fontSize: 60,
+      color: 'blue',
     },
   },
   sizes: {
@@ -43,6 +50,10 @@ const theme = {
       shadowRadius: 14,
       elevation: 25,
     },
+  },
+  linearGradients: {
+    strong: ['primary', 'secondary'],
+    light: ['red', 'green'],
   },
 }
 
@@ -76,7 +87,7 @@ const ResponsiveSquare = () => {
 
 export default function App() {
   return (
-    <DripsyProvider theme={(theme as unknown) as Theme} options={{ ssr: true }}>
+    <DripsyProvider theme={(theme as unknown) as Theme}>
       <Container>
         <View
           sx={{
@@ -84,20 +95,38 @@ export default function App() {
             height: [400, 800],
           }}
         >
-          <H4 sx={{ color: 'text', mb: 2, mt: 0, fontSize: [5] }}>Test</H4>
-          {/* <Pressable>
-            {({ pressed }) => (
-              <DripText sx={{ cursor: 'pointer' }}>
-                {pressed ? 'Joi!' : 'Press Me'}
-              </DripText>
-            )}
-          </Pressable> */}
+          <H4
+            variants={['secondary']}
+            sx={{ color: 'text', mb: 2, mt: 0, fontSize: [5] }}
+          >
+            Test
+          </H4>
           <G variant="primary">Hey</G>
           <G>Hi!</G>
           <View sx={{ bg: 'white', boxShadow: 'md' }}>
             <Text>Card</Text>
           </View>
           <ResponsiveSquare />
+          <Gradient
+            sx={{ height: 50, width: 50, my: 16 }}
+            colors={['primary', 'secondary', '#234fdf']}
+          />
+          <Pressable
+            sx={{
+              height: 50,
+            }}
+          >
+            {({ pressed }) => (
+              <View
+                sx={{
+                  flex: 1,
+                  backgroundColor: pressed ? 'green' : 'red',
+                }}
+              >
+                <Text>You can press me!</Text>
+              </View>
+            )}
+          </Pressable>
         </View>
       </Container>
     </DripsyProvider>