+++ packages/core/src/css/index.tsx
@@ -2,17 +2,16 @@
 import {
   ThemeUIStyleObject,
   CSSObject,
-  ThemeDerivedStyles,
+  UseThemeFunction,
   get,
   Theme,
 } from '@theme-ui/css'
-import { ThemeProvider, SxProp, useThemeUI } from '@theme-ui/core'
-import { useEffect, useRef, useState } from 'react'
-import { Dimensions, Platform, StyleSheet, ScaledSize } from 'react-native'
-// import { useDimensions } from '@react-native-community/hooks'
+import { ThemeProvider, SxProps } from '@theme-ui/core'
+import { Platform, StyleSheet } from 'react-native'
 import type { ThemedOptions, StyledProps } from './types'
 import { defaultBreakpoints } from './breakpoints'
 import type { DripsyTheme } from '../utils/types'
+import { SUPPORT_FRESNEL_SSR } from '../utils/deprecated-ssr'
 
 export { ThemeProvider }
 
@@ -30,14 +29,14 @@ const defaultTheme = {
 
 export type ResponsiveSSRStyles = Exclude<
   ThemeUIStyleObject,
-  ThemeDerivedStyles
+  UseThemeFunction
 >[]
 
 const responsive = (
-  styles: Exclude<ThemeUIStyleObject, ThemeDerivedStyles>,
+  styles: Exclude<ThemeUIStyleObject, UseThemeFunction>,
   { breakpoint }: { breakpoint?: number } = {}
 ) => (theme?: Theme) => {
-  const next: Exclude<ThemeUIStyleObject, ThemeDerivedStyles> & {
+  const next: Exclude<ThemeUIStyleObject, UseThemeFunction> & {
     responsiveSSRStyles?: ResponsiveSSRStyles
   } = {}
 
@@ -59,13 +58,9 @@ const responsive = (
       continue
     }
 
-    if (Platform.OS === 'web') {
+    if (Platform.OS === 'web' && SUPPORT_FRESNEL_SSR) {
       next.responsiveSSRStyles = next.responsiveSSRStyles || []
 
-      // fixed breakpoints for now, not customizable
-      // const breakpoints =
-      //   (Array.isArray(theme?.breakpoints) && theme?.breakpoints) ||
-      //   defaultBreakpoints
       const mediaQueries = [0, ...defaultBreakpoints]
 
       for (let i = 0; i < mediaQueries.length; i++) {
@@ -283,6 +278,21 @@ export const scales = {
 } as const
 type Scales = typeof scales
 
+const positiveOrNegative = (scale: object, value: string | number) => {
+  if (typeof value !== 'number' || value >= 0) {
+    if (typeof value === 'string' && value.startsWith('-')) {
+      const valueWithoutMinus = value.substring(1)
+      const n = get(scale, valueWithoutMinus, valueWithoutMinus)
+      return `-${n}`
+    }
+    return get(scale, value, value)
+  }
+  const absolute = Math.abs(value)
+  const n = get(scale, absolute, absolute)
+  if (typeof n === 'string') return '-' + n
+  return Number(n) * -1
+}
+
 const transforms = [
   'margin',
   'marginTop',
@@ -309,30 +319,12 @@ const transforms = [
   {}
 )
 
-const positiveOrNegative = (
-  scale: Record<string, unknown>,
-  value: string | number
-) => {
-  if (typeof value !== 'number' || value >= 0) {
-    if (typeof value === 'string' && value.startsWith('-')) {
-      const valueWithoutMinus = value.substring(1)
-      const n = get(scale, valueWithoutMinus, valueWithoutMinus)
-      return `-${n}`
-    }
-    return get(scale, value, value)
-  }
-  const absolute = Math.abs(value)
-  const n = get(scale, absolute, absolute)
-  if (typeof n === 'string') return '-' + n
-  return Number(n) * -1
-}
-
 /**
- * Here we remove web style keys from components to prevent annoying errors
+ * Here we remove web style keys from components to prevent annoying errors on native
  */
 const filterWebStyleKeys = (
-  styleProp: Exclude<ThemeUIStyleObject, ThemeDerivedStyles> = {}
-): Exclude<ThemeUIStyleObject, ThemeDerivedStyles> => {
+  styleProp: Exclude<ThemeUIStyleObject, UseThemeFunction> = {}
+): Exclude<ThemeUIStyleObject, UseThemeFunction> => {
   if (Platform.OS === 'web') {
     return styleProp
   }
@@ -340,12 +332,21 @@ const filterWebStyleKeys = (
   // avoid prop mutations
   const finalStyles = { ...styleProp }
   const webOnlyKeys = [
-    'animationKeyFrames',
+    // from https://necolas.github.io/react-native-web/docs/styling/#non-standard-properties
+    'animationKeyframes',
+    'animationFillMode',
     'transitionProperty',
     'whiteSpace',
     'userSelect',
     'transitionDuration',
     'transitionTimingFunction',
+    'cursor',
+    'animationDuration',
+    'animationDelay',
+    'transitionDelay',
+    'animationDirection',
+    'animationIterationCount',
+    'outlineColor',
   ]
   webOnlyKeys.forEach((key) => {
     if (finalStyles?.[key as keyof typeof styleProp]) {
@@ -373,7 +374,8 @@ export const css = (
   const styles = responsive(filteredOutWebKeys, { breakpoint })(theme)
 
   for (const key in styles) {
-    const x = styles[key as keyof typeof styles]
+    // @ts-ignore
+    const x = styles[key]
     const val = typeof x === 'function' ? x(theme) : x
 
     if (key === 'variant') {
@@ -382,6 +384,9 @@ export const css = (
       continue
     }
 
+    /**
+     * @deprecated
+     */
     if (key === 'responsiveSSRStyles' && styles.responsiveSSRStyles) {
       result.responsiveSSRStyles = styles.responsiveSSRStyles.map(
         // here we extract theme values for each item
@@ -401,8 +406,14 @@ export const css = (
       continue
     }
 
+    if (typeof val === 'boolean') {
+      // StyleSheet doesn't allow booleans
+      continue
+    }
+
     const prop = key in aliases ? aliases[key as keyof Aliases] : key
     const scaleName = prop in scales ? scales[prop as keyof Scales] : undefined
+    // @ts-ignore
     const scale = get(theme, scaleName, get(theme, prop, {}))
     const transform = get(transforms, prop, get)
     const value = transform(scale, val, val)
@@ -499,211 +510,10 @@ export const css = (
   return result
 }
 
-// TODO: Do we need options?
-type DefaultOptions = {
-  /**
-   * @default `0`.
-   *
-   * Pass an optional index as the first one. This is useful if you think you know what device users will be on.
-   */
-  defaultIndex?: number
-  /**
-   * You're safe to ignore this. It's for internal use.
-   *
-   * ## Why?
-   *
-   * Since we don't use the RN `Dimensions` API hook on web to determine styles, we need the option to disable this listener on web.
-   */
-  __shouldDisableListenerOnWeb?: boolean
-}
-
-export const useBreakpointIndex = ({
-  defaultIndex = 0,
-  __shouldDisableListenerOnWeb = false,
-}: DefaultOptions = {}) => {
-  // const { width = 0 } = useDimensions().window
-
-  // const getIndex = useCallback(() => {
-  //   // return 1;
-  //   // const { width = 700 } = Dimensions.get("window");
-  //   const breakpointPixels = [...defaultBreakpoints]
-  //     .reverse()
-  //     .find(breakpoint => width >= breakpoint)
-
-  //   let breakpoint = defaultBreakpoints.findIndex(
-  //     breakpoint => breakpointPixels === breakpoint
-  //   )
-  //   breakpoint = breakpoint === -1 ? 0 : breakpoint + 1
-  //   return breakpoint
-  // }, [width])
-
-  const [index, setIndex] = useState(defaultIndex)
-
-  const indexRef = useRef(index)
-
-  useEffect(() => {
-    indexRef.current = index
-  }, [index])
-
-  useEffect(() => {
-    const shouldDisableListener =
-      Platform.OS === 'web' && __shouldDisableListenerOnWeb
-
-    const onChange = ({
-      window: { width },
-    }: {
-      window: ScaledSize
-      screen: ScaledSize
-    }) => {
-      const breakpointPixels = [...defaultBreakpoints]
-        .reverse()
-        .find((breakpoint) => width >= breakpoint)
-
-      let breakpointIndex = defaultBreakpoints.findIndex(
-        (breakpoint) => breakpointPixels === breakpoint
-      )
-      breakpointIndex = breakpointIndex === -1 ? 0 : breakpointIndex + 1
-      if (breakpointIndex !== indexRef.current) {
-        setIndex(breakpointIndex)
-      }
-      // return breakpoint
-    }
-    if (!shouldDisableListener) {
-      Dimensions.addEventListener('change', onChange)
-      onChange({
-        window: Dimensions.get('window'),
-        screen: Dimensions.get('screen'),
-      })
-    }
-    return () => {
-      if (!shouldDisableListener) {
-        Dimensions.removeEventListener('change', onChange)
-      }
-    }
-  }, [__shouldDisableListenerOnWeb])
-
-  return index
-  // return getIndex()
-}
-
-type Values<T> = ((theme: Theme | null) => T[]) | T[]
-
-export function useResponsiveValue<T>(
-  values: Values<T>
-  // options: defaultOptions = {}
-): T {
-  const { theme } = useThemeUI()
-  const array = typeof values === 'function' ? values(theme) : values
-  const index = useBreakpointIndex()
-  return array[index >= array.length ? array.length - 1 : index]
-}
-
-/**
- * This hook is useful when you need to know the responsive value from an array, but on the fly, rather than by rerendering your component.
- */
-// export function useStaticResponsiveValue() {
-//   const { theme } = useThemeUI()
-
-//   return useCallback(
-//     <T extends any>(values: Values<T>) => {
-//       const { width } = Dimensions.get('window')
-//       const getBreakpointIndex = () => {
-//         const breakpointPixels = [...defaultBreakpoints]
-//           .reverse()
-//           .find(breakpoint => width >= breakpoint)
-
-//         let breakpointIndex = defaultBreakpoints.findIndex(
-//           breakpoint => breakpointPixels === breakpoint
-//         )
-//         breakpointIndex = breakpointIndex === -1 ? 0 : breakpointIndex + 1
-//         return breakpointIndex
-//       }
-//       const array = typeof values === 'function' ? values(theme) : values
-//       const index = getBreakpointIndex()
-//       return array[index >= array.length ? array.length - 1 : index]
-//     },
-//     [theme]
-//   )
-// }
-
-export function mapPropsToStyledComponent<P, T>(
-  props: StyledProps<P>,
-  options: ThemedOptions<T>
-) {
-  const {
-    themeKey,
-    defaultStyle,
-    defaultVariant = 'primary',
-    defaultVariants = [],
-  } = options
-  const {
-    breakpoint,
-    sx,
-    theme,
-    variant = defaultVariant,
-    style,
-    variants,
-  } = props
-
-  // overrride the defaults with added ones; don't get rid of them altogether
-  let multipleVariants = [...defaultVariants]
-  if (variants?.length) {
-    multipleVariants = [...defaultVariants, ...variants]
-  }
-  multipleVariants = multipleVariants.filter(Boolean)
-
-  const variantStyle = css(
-    get(theme, themeKey + '.' + variant, get(theme, variant)),
-    breakpoint
-  )({ theme })
-
-  // get the font-family from the variant, and pass it to the other styles as a fallback.
-  // this lets us support customFonts/font weights (https://github.com/nandorojo/dripsy/issues/51)
-  const { fontFamily } = variantStyle
-
-  const baseStyle = css(defaultStyle, breakpoint)({ theme, fontFamily })
-
-  const multipleVariantsStyle = multipleVariants
-    .map((variantKey) =>
-      css(
-        get(theme, themeKey + '.' + variantKey, get(theme, variantKey)),
-        breakpoint
-      )({ theme, fontFamily })
-    )
-    .reduce(
-      (prev = {}, next = {}) => ({
-        ...prev,
-        ...next,
-      }),
-      {}
-    )
-
-  const nativeStyles = css(
-    Array.isArray(style)
-      ? StyleSheet.flatten(style)
-      : StyleSheet.flatten([style]),
-    breakpoint
-  )({ theme, fontFamily })
-
-  const superStyle = css(sx, breakpoint)({ theme, fontFamily })
-
-  // TODO optimize with StyleSheet.create()
-  // TODO IMPORTANT deep merge the `responsiveSSRStyles` from each style above!
-  const styles = () => ({
-    ...baseStyle,
-    ...multipleVariantsStyle,
-    ...variantStyle,
-    ...nativeStyles,
-    ...superStyle,
-  })
-
-  return styles
-}
-
 export class Styles {
-  static create<T>(
-    styles: { [key in keyof T]: SxProp['sx'] }
-  ): { [key in keyof T]: SxProp['sx'] } {
+  static create<T extends { [key: string]: NonNullable<SxProps['sx']> }>(
+    styles: T
+  ): T {
     return styles
   }
 }