+++ packages/core/src/css/styled.tsx
@@ -2,8 +2,6 @@ import type { ComponentType } from 'react'
 import { createThemedComponent } from './create-themed-component'
 import type { ThemedOptions } from './types'
 
-// type Props<P> = Omit<StyledProps<P>, 'theme' | 'breakpoint'>
-
 /**
  * `styled` is little more than a recreation of `createThemedComponent`, with a nicer API. It does the same thing, but looks a bit nicer to use and has a clean name 😇
  *
@@ -17,15 +15,17 @@ import type { ThemedOptions } from './types'
  * ```
  *
  */
-export function styled<P, T>(
+export function styled<P>(
   Component: ComponentType<P>,
   {
     themeKey,
     defaultVariant,
-  }: Pick<ThemedOptions<T>, 'themeKey' | 'defaultVariant'> = {}
+  }: Pick<ThemedOptions, 'themeKey' | 'defaultVariant'> = {}
 ) {
-  return (defaultStyle: ThemedOptions<T>['defaultStyle']) => {
-    return createThemedComponent<P, T>(Component, {
+  return function dripsyFactory<Extra>(
+    defaultStyle?: ThemedOptions<Extra>['defaultStyle']
+  ) {
+    return createThemedComponent<P, Extra>(Component, {
       defaultVariant,
       themeKey,
       defaultStyle,