+++ packages/core/src/provider/index.tsx
@@ -1,63 +1,88 @@
-import React, { ComponentProps, useContext } from 'react'
-import { createMedia } from '@artsy/fresnel'
+import React, { ComponentProps, useEffect, useRef, useState } from 'react'
 import { Platform } from 'react-native'
 import { ThemeProvider as ThemeUIProvider } from '@theme-ui/core'
-import { defaultBreakpoints } from '../css/breakpoints'
+import { MediaContextProvider } from './fresnel'
+import { SUPPORT_FRESNEL_SSR } from '../utils/deprecated-ssr'
 
-const {
-  MediaContextProvider,
-  Media: SSRMediaQuery,
-  createMediaStyle,
-} = createMedia({
-  breakpoints: {
-    // temporary breakpoints for testing fresnel, will update this logic once it works
-    // TODO turn this into a function you use when you first create the app
-    // here, you'll be able to define your own breakpoints maybe?
-    '0': 0,
-    '1': defaultBreakpoints[0],
-    '2': defaultBreakpoints[1],
-    '3': defaultBreakpoints[2],
-    '4': defaultBreakpoints[3],
-  },
-})
+type Props = ComponentProps<typeof ThemeUIProvider> &
+  (
+    | {
+        /**
+         * Set this boolean to `true` if you are using SSR on your app.
+         *
+         * Keep in mind, Dripsy doesn't support SSR anymore.
+         *
+         * Enabling this boolean just renders `null` until your app has mounted on web.
+         *
+         * Set this to `true` if you aren't gating your content already, and if you use responsive arrays in your app.
+         */
+        ssr?: false
+        ssrPlaceholder?: never
+      }
+    | {
+        /**
+         * Set this boolean to `true` if you are using SSR on your app.
+         *
+         * Keep in mind, Dripsy doesn't support SSR anymore.
+         *
+         * Enabling this boolean just renders `null` until your app has mounted on web.
+         *
+         * Set this to `true` if you aren't gating your content already, and if you use responsive arrays in your app.
+         */
+        ssr: true
+        /**
+         * When `ssr` is true, your app will return null on the first mount by default.
+         *
+         * You can use the ssrPlaceholder if you'd prefer.
+         *
+         * ```tsx
+         * <DripsyProvider ssr ssrPlaceholder={<LoadingScreen />}>{children}</DripsyProvider>
+         * ```
+         */
+        ssrPlaceholder?: React.ReactNode
+      }
+  )
 
-const ssrStyleReset = createMediaStyle()
+const useIsSSRReady = ({ ssr = false }: Pick<Props, 'ssr'>) => {
+  const [ready, setReady] = useState(() => {
+    if (Platform.OS !== 'web') {
+      return true
+    }
+    if (ssr) {
+      return false
+    }
+    return true
+  })
 
-const SSRStyleReset = () => (
-  <style type="text/css" dangerouslySetInnerHTML={{ __html: ssrStyleReset }} />
-)
+  const isSSR = useRef(ssr)
 
-export { SSRMediaQuery, SSRStyleReset, ssrStyleReset }
+  useEffect(() => {
+    if (isSSR.current) {
+      setReady(true)
+    }
+  }, [])
 
-type DripsyOptions = {
-  ssr?: boolean
+  return {
+    ready,
+  }
 }
 
-type Props = ComponentProps<typeof ThemeUIProvider> & {
-  options?: DripsyOptions
-}
-
-export let dripsyOptions: DripsyOptions = {
-  ssr: false,
-}
-
-export const setDripsyOptions = (options: Partial<DripsyOptions>) => {
-  dripsyOptions = { ...dripsyOptions, ...options }
-}
-
-const DripsyContext = React.createContext<DripsyOptions>(dripsyOptions)
-
-export function DripsyProvider({ options, ...props }: Props) {
-  if (options) setDripsyOptions(options)
+export function DripsyProvider(p: Props) {
+  const { ssrPlaceholder = null, ssr, ...props } = p
+  const { ready } = useIsSSRReady({ ssr })
 
   const ResponsiveContextProvider =
-    Platform.OS === 'web' ? MediaContextProvider : React.Fragment
+    Platform.OS === 'web' && SUPPORT_FRESNEL_SSR
+      ? MediaContextProvider
+      : React.Fragment
+
+  if (!ready) {
+    return <>{ssrPlaceholder}</>
+  }
 
   return (
     <ResponsiveContextProvider>
-      <DripsyContext.Provider value={{ ssr: !!options?.ssr }}>
-        <ThemeUIProvider {...props} />
-      </DripsyContext.Provider>
+      <ThemeUIProvider {...props} />
     </ResponsiveContextProvider>
   )
 }
@@ -65,11 +90,8 @@ export function DripsyProvider({ options, ...props }: Props) {
 /**
  * @deprecated use DripsyProvider instead
  */
-export function ThemeProvider(props: ComponentProps<typeof DripsyProvider>) {
-  console.warn(
-    '🍷 [Dripsy] ThemeProvider is deprecated, please use DripsyProvider instead. ThemeProvider will be removed next release.'
+export function ThemeProvider() {
+  throw new Error(
+    '🍷 [Dripsy] ThemeProvider is deprecated, please use DripsyProvider instead.'
   )
-  return <DripsyProvider {...props} />
 }
-
-export const useIsSSR = () => !!useContext(DripsyContext).ssr