+++ examples/next-example/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/nandorojo/dripsy/compare/v2.0.1...v2.1.0) (2021-05-21)
+
+**Note:** Version bump only for package next-dripsy-example
+
+
+
+
+
+## [2.0.1](https://github.com/nandorojo/dripsy/compare/v2.0.0-alpha.0...v2.0.1) (2021-05-20)
+
+**Note:** Version bump only for package next-dripsy-example
+
+
+
+
+
+# [2.0.0-alpha.0](https://github.com/nandorojo/dripsy/compare/v1.5.18...v2.0.0-alpha.0) (2021-05-19)
+
+**Note:** Version bump only for package next-dripsy-example
+
+
+
+
+
+## [1.5.8](https://github.com/nandorojo/dripsy/compare/v1.5.7...v1.5.8) (2021-04-20)
+
+**Note:** Version bump only for package next-dripsy-example
+
+
+
+
+
+## [1.5.7](https://github.com/nandorojo/dripsy/compare/v1.5.5-alpha.6...v1.5.7) (2021-04-20)
+
+**Note:** Version bump only for package next-dripsy-example
+
+
+
+
+
+## [1.5.5](https://github.com/nandorojo/dripsy/compare/v1.5.4...v1.5.5) (2021-04-07)
+
+**Note:** Version bump only for package next-dripsy-example
+
+
+
+
+
+## [1.5.4](https://github.com/nandorojo/dripsy/compare/v1.5.3...v1.5.4) (2021-04-07)
+
+**Note:** Version bump only for package next-dripsy-example
+
+
+
+
+
 ## [1.5.3](https://github.com/nandorojo/dripsy/compare/v1.5.2...v1.5.3) (2021-03-28)
 
 **Note:** Version bump only for package next-dripsy-example